+++ src/main/java/lotto/Lotto.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import utils.LottoValidator;
+
 import java.util.ArrayList;
 import java.util.List;
