+++ src/main/java/utils/LottoCreator.java
@@ -1,6 +1,7 @@
-package lotto;
+package utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.Lotto;
 
 import java.util.Collections;
 import java.util.List;
